Update typemap for Duckdb v0.10.1 (#28)
* fix: update typemaps to duckdb v0.10.1

fix: update typemaps to duckdb v0.10.1

This commit reverts the breaking type mapping definitions that the
duckdb team introduced in `v0.10.0` and switches them to how they do
them in `v0.10.1` / `v0.9.2`. ie. They realized that they broke all
C-based APIs and undid the UHUGEINT insertion.

In here is also a bit of a refactor that tries to get some of the
DRYness of the `define-long-enums` while also mimicking the new style of
the C api (where they use explicit mappings on enums). See [this pull
request](https://github.com/duckdb/duckdb/pull/10649) for details. I
think this strikes a nice balance between clarity and DRYness.

Lastly we now only error on version `v0.10.0` since this should restore
compatibility with `v0.9.2`. I confirmed that the test suite passes when
using duckdb `v0.9.2` via nix flakes.

* chore: add duckdb to nix flake devShell

Add duckdb to the devShell in the nix flake so that we can easily test
different versions of Duckdb

* chore: update nix flake to duckdb v0.10.1

Get the latest version of duckdb