fix: pivot table and multi-statement support (#31)
* chore: update nix flake build

Updates the nix flake build to handle the new multi-output format of the
duckdb library.

* fix: pivot table and multi-statement support

This commit introduces support for pivot tables by adding support for
multiple statements inside a prepared statement. This is because a pivot
table will automatically create an anonymous type for the result before
running.

This commit includes several refactors to help cleanup the code.

1) Updates the duckdb-ffi function defenitions by changing `:pointer` to
`by-value :type` defenitions. This allows the use of resource tracking
on the structs to call the cleanup functions, as well as allowing the
structs to be passed as `:pointer` when approprirate (avoiding a lot of
the manual `Pointer.` construction) via the auto casting of `->pointer`.

2) Get rid of the multiple types for `PrepStatement` arities and instead
just defenining multiple `IFn` impls for the various arities.

3) Get rid of the `param-types` vector which wasn't being used in the
current version of param binding.

4) Get rid of the need to destroy resources manually and instead rely on the resource tracking.

5) Introduce a `->error` fn helper for generating error messages from
c-string ptrs since we had a lot of code that was doing the required
ceremony inline.

6) Introduce a `-execute-stmt!` helper that handles resource binding and
cleanup to make it easier to understand the logic inside `prepare!`